[195063] - rewrite backend to go reverse proxy for forum and core apis
