+++ .gitignore
@@ -419,3 +419,5 @@ package
 .ionide
 
 # End of https://www.toptal.com/developers/gitignore/api/intellij,intellij+iml,intellij+all,visualstudiocode,go,node,svelte
+
+code-help-admin-backend/admin-api/cmd/cmd
