--- code-help-admin-backend/@global/deploy/docker-compose.yaml
@@ -1,6 +1,0 @@
-services:
-  hello:
-    container_name: hello-container
-    image: hello:latest
-    ports:
-      - "5080:8080"
