--- code-help-admin-backend/admin-api/go.mod
@@ -1,24 +1,0 @@
-module admin-api
-
-go 1.22
-
-require (
-	github.com/coreos/go-oidc/v3 v3.10.0
-	github.com/labstack/gommon v0.4.2
-	github.com/oapi-codegen/runtime v1.1.1
-)
-
-require (
-	github.com/go-jose/go-jose/v4 v4.0.1 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/google/uuid v1.5.0 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasttemplate v1.2.2 // indirect
-	golang.org/x/crypto v0.19.0 // indirect
-	golang.org/x/oauth2 v0.13.0 // indirect
-	golang.org/x/sys v0.17.0 // indirect
-	google.golang.org/appengine v1.6.8 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
