--- code-help-admin-backend/admin-api/internal/admin/forum/community/mapper.go
@@ -1,1 +1,0 @@
-package community
