--- code-help-admin-backend/api-spec/go.mod
@@ -1,21 +1,0 @@
-module api-spec
-
-go 1.22
-
-require (
-	github.com/getkin/kin-openapi v0.123.0
-	github.com/oapi-codegen/runtime v1.1.1
-)
-
-require (
-	github.com/apapsch/go-jsonmerge/v2 v2.0.0 // indirect
-	github.com/go-openapi/jsonpointer v0.20.2 // indirect
-	github.com/go-openapi/swag v0.22.8 // indirect
-	github.com/google/uuid v1.5.0 // indirect
-	github.com/invopop/yaml v0.2.0 // indirect
-	github.com/josharian/intern v1.0.0 // indirect
-	github.com/mailru/easyjson v0.7.7 // indirect
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
-	github.com/perimeterx/marshmallow v1.1.5 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
