--- code-help-admin-backend/go.work
@@ -1,6 +1,0 @@
-go 1.22
-
-use (
-	admin-api
-	api-spec
-)
